+++ module-check/src/main/clojure/clojure/core/typed/check/print_env.clj
@@ -27,7 +27,7 @@
                            [k (prs/unparse-type v)]))
            :props (map prs/unparse-filter (:props e))
            :aliases (:aliases e)
-           ;:frees (map (t/fn> 
+           ;:frees (map (t/fn
            ;              [nme :- t/Sym, bnd :- (U nil Bounds)]
            ;              {:pre [(symbol? nme)
            ;                     ((some-fn nil? r/Bounds?) bnd)]}