+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Breaking: Removed deprecated/unsupported macros to improve loading times
   [x] clojure.core.typed/dotimes>
-  [ ] clojure.core.typed/for>
+  [x] clojure.core.typed/for>
   [ ] clojure.core.typed/doseq>
   [ ] clojure.core.typed/fn>
   [ ] clojure.core.typed/def>