+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   [ ] clojure.core.typed/fn> (use t/fn)
   [x] clojure.core.typed/defn> (use t/defn)
   [x] clojure.core.typed/def> (use t/defn)
-  [ ] clojure.core.typed/defprotocol> (use t/defprotocol)
+  [x] clojure.core.typed/defprotocol> (use t/defprotocol)
   [x] clojure.core.typed/loop> (use t/loop)
   [x] clojure.core.typed/atom> (use t/atom)
   [x] clojure.core.typed/ref> (use t/ref)