+++ CHANGELOG.md
@@ -14,7 +14,7 @@
   [x] clojure.core.typed/ref> (use t/ref)
   [x] clojure.core.typed/def-alias (use t/defalias)
   [x] clojure.core.typed/ann-precord (use t/ann-record)
-  [ ] clojure.core.typed/ann-pprotocol (use t/ann-protocol)
+  [x] clojure.core.typed/ann-pprotocol (use t/ann-protocol)
 
 # 0.5.4 - 13 September 2018