+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.6.0 - SNAPSHOT
+# 0.6.0 - 8 November 2018
 
 - Breaking: Removed deprecated/replaced macros (since 0.2.45) to improve future loading times
   - clojure.core.typed/dotimes> (use t/dotimes)